Refactor preferences for cuda standalone device
closes #195, closes #140
It also enables logging of cuda preferences.
