+++ brian2cuda/tests/test_cuda_generator.py
@@ -370,13 +370,13 @@ def test_default_function_convertion_preference():
 
     unrepresentable_int = 2**24 + 1  # can't be represented as 32bit float
 
-    prefs.codegen.generators.cuda.default_functions_integral_convertion = np.float32
+    prefs.devices.cuda_standalone.default_functions_integral_convertion = np.float32
     G = NeuronGroup(1, 'v: 1')
     G.variables.add_array('myarr', dtype=np.int32, size=1)
     G.variables['myarr'].set_value(unrepresentable_int)
     G.v = 'floor(myarr)'.format(unrepresentable_int)
 
-    prefs.codegen.generators.cuda.default_functions_integral_convertion = np.float64
+    prefs.devices.cuda_standalone.default_functions_integral_convertion = np.float64
     G2 = NeuronGroup(1, 'v: 1')
     G2.variables.add_array('myarr', dtype=np.int32, size=1)
     G2.variables['myarr'].set_value(unrepresentable_int)
@@ -422,7 +422,7 @@ def test_default_function_convertion_warnings():
         G4.variables.add_array('myarr', dtype=np.uint32, size=1)
         G4.v = 'arcsin(i*myarr)'
 
-    prefs.codegen.generators.cuda.default_functions_integral_convertion = np.float32
+    prefs.devices.cuda_standalone.default_functions_integral_convertion = np.float32
 
     BrianLogger._log_messages.clear()
     with catch_logs() as logs5:
