+++ brian2cuda/tools/utils.py
@@ -122,7 +122,7 @@ def set_preferences(args, prefs, fast_compilation=True, suppress_warnings=True,
     if suppress_warnings:
         # Surpress some warnings from nvcc compiler
         compile_args = ['-Xcudafe "--diag_suppress=declared_but_not_referenced"']
-        prefs['codegen.cuda.extra_compile_args_nvcc'].extend(compile_args)
+        prefs['devices.cuda_standalone.cuda_backend.extra_compile_args_nvcc'].extend(compile_args)
         prints.append("Suppressing compiler warnings")
 
     if args.jobs is not None:
@@ -149,7 +149,7 @@ def set_preferences(args, prefs, fast_compilation=True, suppress_warnings=True,
 
         if not args.no_atomics:
             all_prefs_list.append(
-                {'codegen.generators.cuda.use_atomics': False})
+                {'devices.cuda_standalone.use_atomics': False})
 
         # create a powerset (all combinations) of the all_prefs_list
         all_prefs_combinations = []
@@ -168,7 +168,7 @@ def set_preferences(args, prefs, fast_compilation=True, suppress_warnings=True,
         fixed_prefs_dict[k] = v
 
     if args.no_atomics:
-        k = 'codegen.generators.cuda.use_atomics'
+        k = 'devices.cuda_standalone.use_atomics'
         v = False
         prefs[k] = v
         fixed_prefs_dict[k] = v
