+++ dev/benchmarks/run_manunscript_runtime_vs_N_benchmarks.py
@@ -134,11 +134,11 @@ configs = [# configuration                          project_directory
 
 
           #(DynamicConfigCreator('CUDA standalone (max blocks, no atomics)',
-          #                      prefs={'codegen.generators.cuda.use_atomics': False}),
+          #                      prefs={'devices.cuda_standalone.use_atomics': False}),
           # 'cuda_standalone'),
 
           #(DynamicConfigCreator('CUDA standalone (1 block, no atomics)',
-          #                      prefs={'codegen.generators.cuda.use_atomics': False,
+          #                      prefs={'devices.cuda_standalone.use_atomics': False,
           #                             'devices.cuda_standalone.parallel_blocks': 1}),
           # 'cuda_standalone'),
